Move duplicate recipe preview code to AbstractRecipePreviewMenu
